+++ caliban-v2x-plugin/src/main/resources/skywalking-plugin.def
@@ -1,1 +1,1 @@
-caliban-2.x=org.bitlap.skywalking.apm.plugin.caliban.define.CalibanInstrumentation
+caliban-2.x=org.bitlap.skywalking.apm.plugin.caliban.v2.define.CalibanInstrumentation
