+++ caliban-v2x-plugin/src/main/scala/org/bitlap/skywalking/apm/plugin/caliban/v2/CalibanInterceptor.scala
@@ -1,4 +1,4 @@
-package org.bitlap.skywalking.apm.plugin.caliban
+package org.bitlap.skywalking.apm.plugin.caliban.v2
 
 import java.lang.reflect.Method
